Use a single shared_memory_object
Memory and Info had two separate objects previously under
two different names: "<topic>Mem" and "<topic>Info", but
that made the management of shared memory cumbersome with
duplication of code. Replace that with using a single
shared memory object with pointer offsets for the Memory
and Info objects, in the future we'll incorporate the
dynamic buffer in this memory space as well.