FIX: Replace raw pointer with unique_ptr arrays
In the raw bytes usage of subscriber, the buffer was malloc'd
while reading from the shared memory. This heap space had to be
manually free'd by the user in the callback.

Handling the free'ing in the case of an exception was also a little
messy requiring two free's, one in the try and another in the catch.