Remove queue_size template parameter
Memory could have a variable sized underlying circular buffer,
but this parameter had to passed as a template parameter during
compile time. This meant that most of the implementation was
restricted to the header files.

Removing this makes the implementation of the end-points for both
PubSub/RPC much nicer, in that, it requires no template parameters.
Instead of variable-sized circular buffer, we stick to a fixed
size of *128*. It handle recency constraints, I can introduce a
"Policy" type. Or re-visit runtime parameter for Memory.