Remove queue_size template parameter (#40)
Memory could have a variable-sized underlying circular buffer, but the size had
to be passed as a template parameter during compile time. This meant that
most of the implementation was restricted to header files.

Removing this makes the implementation of the end-points for both PubSub/RPC
much nicer; in that, it requires no template parameters. Instead of the variable-
sized circular buffer, we stick to a fixed size of 128. It handles recency constraints;
I can introduce a "Policy" type. Or re-visit runtime parameter for Memory.

ALSO: Refactor Topic and Channel to use Memory via composition instead of
inheritance. I have no idea why I went with inheritance in the first place.