[pubsub] Fix lag jumpahead logic
The existing behavior when the subscriber lags (behind the write head by
queue size, since the buffer is circular, new data will have been written
to the location the subscriber is trying to read from) was to move to the
oldest valid location. However, since the jumpahead was happening outside
of any locks, between the time of jumping ahead and acquring the read lock
for that location, the publisher(s) could have written past the new read
location.

The fix was to jump ahead optimistically, rather than to the last valid
location. Currently, this is mid-way between the write head and the last
valid location. Another issue that surfaced was that since the jumpahead
happens outside any lock, in the case of particularly fast publishers, even
here the same issue can be seen, where the write head wraps around and writes
past the new (optimistic) read location.

The second fix was to do a similar jump ahead (and associated check) while
under read lock. This is an expensive operation, since we might potentially
by readig from two buffer elements. However, this is a (very rate) slow path.

To ensure this logic is correct, the benchmark for pubsub has an assert to
check that each subsequent subscribed elements is increasing. The benchmark
has also been rewritten to make it much cleaner.